Implement JPA DAO

+++ src/main/java/com/example/CourseWorkWithDB/Validators/NumberValidator.java
@@ -4,7 +4,7 @@ import com.example.CourseWorkWithDB.Exceptions.LessThanGivenException;
 import com.example.CourseWorkWithDB.Exceptions.LessThanZeroException;
 
 public class NumberValidator {
-    public static void moreThanZero(long value, String valueName) throws LessThanZeroException {
+    public static void moreThanZero(double value, String valueName) throws LessThanZeroException {
         if(value<0) throw new LessThanZeroException(valueName + " should be more than or equal to zero");
     }
